requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
automium==0.2.6
automium-web==0.1.1
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-admin-black==1.0.14
django-allauth==65.1.0
django-autoslug==1.9.9
django-cache==0.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-grappelli==4.0.1
django-js-asset==2.2.0
django-material==1.12.0
django-redis==5.4.0
django-theme-material-kit==1.0.25
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
easy-thumbnails==2.10
exceptiongroup==1.2.2
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
legacy==0.1.7
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.22.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
views==0.3
websocket-client==1.8.0
Werkzeug==3.0.4
whitenoise==6.8.2
wsproto==1.2.0
